synesthesia-core/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "synesthesia"
version = "0.1.0"
description = "Real-Time Latent Space Audio Navigation System"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "numpy==1.26.4",
    "torch==2.2.0",
    "transformers==4.37.2",
    "qdrant-client",
    "python-dotenv",
    "pydub",
    "spotipy",
    "python-multipart"
]

[tool.maturin]
python-source = "python"
module-name = "synesthesia.core"
